refactor: update mean finance to balmy
