--- coins/src/adapters/yield/mean-finance/index.ts
@@ -1,13 +1,0 @@
-import getTokenPrices from "./mean-finance";
-
-export function meanFinance(timestamp: number = 0) {
-  return Promise.all([
-    getTokenPrices("optimism", timestamp),
-    getTokenPrices("polygon", timestamp),
-    getTokenPrices("arbitrum", timestamp),
-    getTokenPrices("ethereum", timestamp),
-    getTokenPrices("bsc", timestamp),
-    getTokenPrices("xdai", timestamp),
-    getTokenPrices("moonbeam", timestamp),
-  ]);
-}
