+++ coins/src/adapters/yield/misc4626/index.ts
@@ -2,7 +2,7 @@ import getTokenPrices from "./misc";
 import tokens from "./tokens.json";
 import tokensQiDAO from "./tokensQiDAO.json";
 import { getYieldWrites2 } from "../../utils/yieldTokens";
-import { unwrap4626 } from "../mean-finance/mean-finance";
+import { unwrap4626 } from "../balmy/balmy";
 
 export function misc4626(timestamp: number = 0) {
   const calls = Object.keys(tokens).map((c) => getTokenPrices(c, timestamp))
